+++ CHANGELOG.md
@@ -1,14 +1,16 @@
 <!-- markdownlint-disable first-line-h1 -->
 ## v1.32.0
 
-*2025-12-02*
+*2025-12-04*
 
 **Added**
 
 - HTML conversion of `<syntaxhighlight>` with [Prism](https://prismjs.com/) highlighting
+- [`CategoryToken.innerText`](https://github.com/bhsd-harry/wikiparser-node/wiki/CategoryToken-%28EN%29#innertext)
 - [`Token.prototype.toHtml`](https://github.com/bhsd-harry/wikiparser-node/wiki/Token-%28EN#tohtml) now supports category links at the end of the page
 - [`Parser.callParserFunction`](https://github.com/bhsd-harry/wikiparser-node/wiki/Parser-%28EN%29#callparserfunction) now supports custom parser functions defined via [`Parser.setFunctionHook`](https://github.com/bhsd-harry/wikiparser-node/wiki/Parser-%28EN%29#setfunctionhook)
 - [`Title.prototype.getFileUrl`](https://github.com/bhsd-harry/wikiparser-node/wiki/Title-%28EN%29#getfileurl) that internally calls the `filepath` parser function to generate a `Special:Redirect/file/` URL
+- [`Parser.lint`](https://github.com/bhsd-harry/wikiparser-node/wiki/Parser-%28EN%29#lint), [`Parser.print`](https://github.com/bhsd-harry/wikiparser-node/wiki/Parser-%28EN%29#print) and [`Parser.toHtml`](https://github.com/bhsd-harry/wikiparser-node/wiki/Parser-%28EN%29#tohtml)
 
 **Fixed**